cscli: extract package 'crowdsec-cli/clientinfo'
